+++ lib/viewmodels/diary/diary_notifier.dart
