+++ lib/views/diary/diary_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mindow/viewmodels/providers/diary_notifier.dart';
+import 'package:mindow/viewmodels/diary/diary_notifier.dart';
 import 'package:mindow/views/diary/widgets/diary_date.dart';
 import 'package:mindow/views/diary/widgets/emotion_grid.dart';
